+++ src/histogram.c
@@ -2,6 +2,11 @@
  * @file histogram.c
  * @brief Implementation of the p99 performance percentile histogram.
  *
+ * Home: https://github.com/synesissoftware/p99
+ *
+ * Created: 4th July 2026
+ * Updated: 6th July 2026
+ *
  * @copyright Copyright (c) 2026, Matthew Wilson and Synesis Information
  *   Systems
  * @license BSD-3-Clause
@@ -295,7 +300,7 @@ p99_histogram_push_event_time_s(
 
 /* --- Statistics ------------------------------------------------------- */
 
-P99_CALL(size_t)
+P99_CALL(uint64_t)
 p99_histogram_event_count(p99_histogram_t const* histogram)
 {
     return histogram->event_count;
@@ -496,7 +501,7 @@ p99_histogram_value_at_p50(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   1
     ,   2
     );
@@ -515,7 +520,7 @@ p99_histogram_value_at_p75(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   3
     ,   4
     );
@@ -534,7 +539,7 @@ p99_histogram_value_at_p90(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   90
     ,   100
     );
@@ -553,7 +558,7 @@ p99_histogram_value_at_p95(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   95
     ,   100
     );
@@ -572,7 +577,7 @@ p99_histogram_value_at_p99(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   99
     ,   100
     );
@@ -591,7 +596,7 @@ p99_histogram_value_at_p99_5(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   995
     ,   1000
     );
@@ -610,7 +615,7 @@ p99_histogram_value_at_p99_9(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   999
     ,   1000
     );
@@ -629,7 +634,7 @@ p99_histogram_value_at_p99_99(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   9999
     ,   10000
     );
@@ -648,7 +653,7 @@ p99_histogram_value_at_p99_999(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   99999
     ,   100000
     );
@@ -667,7 +672,7 @@ p99_histogram_value_at_p99_999_9(
 )
 {
     uint64_t target_rank = p99_u64_mul_div_u64_(
-        (uint64_t)histogram->event_count
+        histogram->event_count
     ,   999999
     ,   1000000
     );
